requirements.txt
# ── Backend API ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# ── LLM & Agents ──
langchain==0.3.14
langgraph==0.2.60
langchain-openai==0.3.0
langchain-anthropic==0.3.0
langchain-community==0.3.0
openai>=1.0.0

# ── Database ──
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6

# ── Optimization ──
ortools==9.11.4210

# ── Utilities ──
python-dotenv==1.0.1
httpx==0.28.1
PyJWT==2.12.1
passlib==1.7.4
bcrypt==5.0.0
neo4j==6.1.0
tavily-python==0.7.3